+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Notable changes to this project.
 
 ## [] -
 - fix downloading dataset for tournaments > 1
+- add `validationCorrelation` and `liveCorrelation` to all relevant places
 
 ## [1.6.0] - 2019-07-10
 - default to tournament 8 `katzuagi`